Give the site a face lift
The menu took up a lot of space on smaller screens and the icons maybe
confusing for new visitors. So the menu is now in the top and uses
plain text links.

Why no example code? I could not come up with any example code that is
short and exciting. I think it would be much better to have in depth
tutorials and/or real programs to show off instead. That however left
the index page empty which allowed the logo to be right in your face
with a description of what the site is about when you land on the
index page.

I simplified the values in the logo svg then inlined it on the index
page. This simplifies the menu and saves a file request.

I put blog entries on the index page as well since there are only
2 at the moment and considering the rate at which entries have been
added it will probably take years before paging is needed.

The previous fonts for text and code clashed in both sizes and families.

Tables were of variable width which I think looked weird. The width of
the table cells are set to *auto* which means that the contents dictate
the width. Seems good for now.

Header margins and weight.
To make headers more distinct, especially in the spec, they are now in
bold, have 15px top margins and are in different sizes proportionally
to their level.

I chose to use Sentence Casing for headers since it doesn't have
different interpretations like Title Casing has.

Like paragraphs and many other elements list now have a 15px top margin.