Define strings and characters
`string_literal`s are UTF-8 encoded and are syntax sugar for `[]U8`.
This means that it is the programmer's job to keep track of the encoding of
each and every string. I suspect that this would mostly be a problem for
programs dealing with many encodings all over the place.

Character literals such as `'A'` are syntactic sugar for the Unicode
code point as a `U32`.

The standard library will have some functions for dealing with UTF-8. These
functions will be prefixed with `utf8_` so they won't clash with 3rd party
functions. Other encodings are not supported.

To begin with I looked at how D and Rust do strings. It is nice that you have
types for specific encodings but you pay for with some complexity. Treating
strings as byte arrays makes that complexity go away, for the language and
compiler. The programmer pays for this, as the compiler won't notice if you are
going to bulldoze strings by mixing encodings or pass a string to a non-string
function.

I'm not sure what a dedicated string type would buy the language besides that
the compiler can point out some type errors. Beyond that, functions would need
to know how the string is encoded by having an enum field which is pretty much
the same as having distinct string types. You could also store the string
data as an `[]U32` which avoids the need for the enum field but at the cost of
memory.

Beyond that, there were some minor things fixed at the same time:
* `string_literal` specified that there could only be one character.
  This was wrong and now there can be zero or more characters in a
  `string_literal`

* Fixed a reference to `string` which should have been `string_literal`.