I messed up the syntax rules for type names and namespaces some time ago and only recently realized this. The primitives of the language are:
* I8
* I16
* I32
* I64
* U8
* U16
* U32
* U64
* F32
* F64
* Bool

but the syntax was changed to
```
upper_case_identifier ← [A-Z][a-z]* (_([A-Z][a-z]*|[0-9]+))* whitespace
```

The nice thing, subjectively speaking, is the separation between numbers and
letters in the name. The `Chip_8` namespace looks kinda nice. However the above
syntax rule do not match the primitives and to be honest `U_32` looks nasty
in my opinion. So to fix this `upper_case_identifier` has been changed back to:

```
upper_case_identifier ← [A-Z][a-z0-9]* (_[A-Z][a-z0-9]*)* whitespace
```

and `lower_case_identifier` changed to:

```
lower_case_identifier ← !keyword [a-z][a-z0-9]* (_[a-z][a-z0-9]*)* whitespace
```