Overhaul expression syntax
From expression to multiplicative_expression was fine but prefix_expression
could only handle a single prefix operator.

postfix_expression did not work at all. Either it ended up as a
primary_expression or as a postfix_expression without a primary_expression.

size_of_expreassion was defined as a prefix operator but acted like
a primary_expression. It is now defined as a primary_expression.

Variable references are now defined.