Make size_of_expression inferable
There is no reason to fix the type of the size_of_expression. While we are
at it, getting the size of an expression's type is also helpful. Taking the
size of an expression in stead of hard coding the type means that the type
can change without silently breaking code.