Remove most type inference
I have decided to make a fairly big change to the language. So far `a = 1` could
be an assignment or variable declaration. So a bit of type inference was
introduced. Basically the left hand side is what ever the type is on the right
hand side.

But, that leaves the question which one of the integer types is it when
declaring the variable? I "solved" that with optional postfixes for integer and
float literals that specified the type.That allowed further assignments to omit
the postfix. However, another and bigger truck was headed my way. Soon after I
thought it would be nice to omit postfixes entirely:

```
function do_stuff
    output I32

    a = 1

    return a
end
```

The type checker needs type variables to be able to figure this out. I thought
"Great! Lets do that!" after I had done some tests implementing it. I did run
into a bit of trouble with compound types but that was quickly resolved.

The real trouble starts when overloaded and or generic functions comes into
the mix because the type checker is then unable to figure out the signature
of `x` by name alone. It has to get context from the expected return type,
the actual arguments to `x` etc. However, `x`s type might be a type variable
or one or more of the actual arguments' types may be a type variable.
While this is not impossible to solve I don't think the amount of
complexity to the spec and compiler is worth not explicitly writing
`I32 a = 42` considering that type inference was only meant to deal with
types for integer and float literals. Compound and array literals where
explicit about types.