Define function calls and other things
Overload resolution based on actual_type_parameters have been specified.

I threw out my own terms for generics stuff in favor of terms that people
actually use. That also removed the `input` and `output` keywords in favor of
a more C like syntax `function a(I32 b) : U8`.

Because of this I also replaced the `of` keyword for specifying the underlying
type of enumerations with `:`.

I also took the opportunity change the syntax of function types so that they
match how signatures look but without the function keyword.

I changed the array literal syntax. I want them to work similarly to how
compound literals work. Inference and all that.

Css - I find it really nice that I can distinguish between type, keywords
and rules in the spec but going back between fonts all the times is anoying
for me to read. So one font to rule them all. I also widened the pages. They
still fit on half the width of 1080p monitors.