Remove propositions and assertions
The bulk of the testing I do is wired up using plain functions so the
built in language support for propositions and assertions would only
handle simple cases but would be inadequate for handling variants of the
same test.

I think the exact testing mechanism is best left to the end user who has
better knowledge of their system than I do. That said there would be a
need for reporting file paths and line numbers in testing. I could take
a page from C and use __FILE__ and __LINE__ to do just that.