C style qualifiers in place of privileges
The more I looked at privileges the more broken they turned out to be!

```
structure Data
    #I32 a
end

extern function a() : #$I32

function b() : I32
    Data d = { a = a() } // Write privilege is lost here...
    #?Data dd = #d       // New read privilege here...

    return @dd.a         // Write only violation here!
end
```

```
extern function a() : #$I32

function b(#$#I32 c)
    @c = a()
end

function d(#`!?$I32 e)
    b(#e) // e == a() with more privileges then intended!
end
```

If only formal parameter types can be noalias qualified you can't assign
the parameter to anything but another formal parameter which is undefined
behavior. So don't do that of course. More importantly it avoids the need
to define what "based on" a pointer means. I stared at section 6.7.3.1 in
the C99 spec and I have no clear understanding of what that really means.

Its one of those things that feels like I almost understand it but I can't
explain how it works. And every time I sit down and read section 6.7.3.1 again
or someone elses interpretation/regurgitation of it I sense a glimmer of
understand but it falls apart immediatly.

I haven't used restrict for anything else than function parameters and can't
recall any code in the wild doing it. Did not find during a very quick search.
So while there probably are use cases for restrict on other things than
function parameters and on expressions "based on" some other pointer I
decided that

If it turns out that it is needed by someone and that someone can ELI5 that
use case and the exact semantics hit me up.

Qualifiers are ignored when checking for matching functions in scope.
I could not come up with a case for overloading on qualifiers. So its
simpler to not doing that. It can be added later if needed.